Add configuration of multipathServiceType in NIOTSConnectionBootstrap
Motivation:

Allow different devices to leverage the capabilities of Mutipath TCP (MPTCP) to
enhance the network reliability. Thanks to MPTCP, a connection can for example
automatically migrate to another interface if it deteriorates on the first one.
This can be especially interesting on MacOS X and iOS, where devices may frequently
benefit from multiple interfaces (ethernet + Wi-Fi for Macs and Wi-fi + cellular for
iOS). Allowing developers to enable MPTCP on their connections seems thus like a fine
addition to this library.

Modifications:

Added a function "withMultipath" on NIOTSConnectionBootstrap, that allow to configure
the type of service used for MPTCP (defaults to disabled). This value will be stored
in a field, and then propagated to the underlying channel when the connect method
will be called. Also updated the parameters field of NIOTSConnectionChannel to set
the multipathServiceType accordingly.

Result:

Users will now be able to easily enable Multipath TCP, allowing them to benefit from
seamless handover, interactive mode to automatically use the lowest delay interface or
aggregate mode to send data in parallel on both interfaces.